requirements.txt
# Snowflake connector
snowflake-connector-python==3.12.2
snowflake-sqlalchemy==1.6.1

# Data analysis and manipulation
pandas==2.2.3
numpy==1.26.4

# Jupyter for interactive analysis
jupyter==1.1.1
ipykernel==6.29.5

# Data visualization
matplotlib==3.9.2
seaborn==0.13.2
plotly==5.24.1

# Environment management
python-dotenv==1.0.1

# SQL utilities
sqlalchemy==2.0.35

